Refactor:팔로우 버튼 공용 컴포넌트로 이동(추천리스터>공용)
