+++ src/app/(home)/_components/RecommendedUsers.css.ts
@@ -78,24 +78,3 @@ export const recommendUserNickname = style({
   fontWeight: 400,
   letterSpacing: '-0.28px',
 });
-
-export const followButtonDefault = style({
-  width: 'auto',
-  height: 'auto',
-  padding: '4px 6px',
-
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-
-  backgroundColor: '#3D95FF',
-  borderRadius: '20px',
-  color: vars.color.white,
-  fontSize: '1.2rem',
-  fontWeight: '400',
-});
-
-export const followButtonFollowing = style({
-  backgroundColor: vars.color.white,
-  color: '#8599AD',
-});
