+++ src/app/(home)/_components/RecommendedUsers.tsx
@@ -7,7 +7,7 @@ import { useQuery } from '@tanstack/react-query';
 import { QUERY_KEYS } from '@/lib/constants/queryKeys';
 import getRecommendedUsers from '@/app/_api/home/getRecommendedUsers';
 import { useUser } from '@/store/useUser';
-import FollowButton from './FollowButton';
+import FollowButton from '@/components/FollowButton/FollowButton';
 import { UserProfileType } from '@/lib/types/userProfileType';
 
 import fallbackProfile from '/public/images/fallback_profileImage.webp';
