+++ src/app/user/locale.ts
@@ -5,16 +5,15 @@ export const userLocale = {
       follower: '아직은 팔로워가 없어요',
       following: '아직 팔로우한 사람이 없어요',
     },
+    follow: '팔로우',
     follower: '팔로워',
     following: '팔로잉',
+    cancelFollow: '팔로우 취소',
     total: '총',
     people: '명',
     myList: '마이 리스트',
     collaboList: '콜라보 리스트',
     noListMessage: '해당 카테고리에 아직 리스트가 없어요',
-    cancelFollow: '팔로우 취소',
-    follow: '팔로우',
-
     confirm: '확인',
     goToMypage: '마이페이지로 이동하기',
     profileImageAlt: '프로필 이미지',
