+++ src/components/FollowButton/FollowButton.tsx
@@ -11,13 +11,14 @@ import { QUERY_KEYS } from '@/lib/constants/queryKeys';
 import { UserType } from '@/lib/types/userProfileType';
 import toasting from '@/lib/utils/toasting';
 import toastMessage, { MAX_FOLLOWING } from '@/lib/constants/toastMessage';
-import * as styles from './RecommendedUsers.css';
 
 import useBooleanOutput from '@/hooks/useBooleanOutput';
 import Modal from '@/components/Modal/Modal';
 import LoginModal from '@/components/login/LoginModal';
 import { useLanguage } from '@/store/useLanguage';
 import { commonLocale } from '@/components/locale';
+
+import * as styles from './FollowButton.css';
 
 interface FollowButtonProps {
   isFollowing: boolean;
